# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Expense Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - name: accountId
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Transaction"
                  total:
                    type: integer
                required:
                  - transactions
                  - total
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionInput"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
    put:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionInput"
      responses:
        "200":
          description: Transaction updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List all notification reminders
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: "#/components/schemas/NotificationSetting"
                required:
                  - notifications
    post:
      operationId: createNotification
      tags: [notifications]
      summary: Create a new notification reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationInput"
      responses:
        "201":
          description: Notification created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSetting"
  /notifications/{id}:
    put:
      operationId: updateNotification
      tags: [notifications]
      summary: Update a notification reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationInput"
      responses:
        "200":
          description: Notification updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSetting"
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete a notification reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
          format: double
        type:
          type: string
          enum: [income, expense]
        description:
          type: string
        merchantName:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
        categoryId:
          type: string
          nullable: true
        categoryName:
          type: string
          nullable: true
        categoryColor:
          type: string
          nullable: true
        categoryIcon:
          type: string
          nullable: true
        accountId:
          type: string
          nullable: true
        accountName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        importSource:
          type: string
          enum: [manual, email]
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - amount
        - type
        - description
        - date
        - createdAt
        - updatedAt
    CreateTransactionInput:
      type: object
      properties:
        amount:
          type: number
          format: double
        type:
          type: string
          enum: [income, expense]
        description:
          type: string
        merchantName:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
        categoryId:
          type: string
          nullable: true
        accountId:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        importSource:
          type: string
          enum: [manual, email]
          nullable: true
      required:
        - amount
        - type
        - description
        - date
    UpdateTransactionInput:
      type: object
      properties:
        amount:
          type: number
          format: double
        type:
          type: string
          enum: [income, expense]
        description:
          type: string
        merchantName:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
        categoryId:
          type: string
          nullable: true
        accountId:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
    NotificationSetting:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [daily_review, weekly_summary, bill_reminder, budget_alert, custom]
        frequency:
          type: string
          enum: [daily, weekly, monthly, once]
        time:
          type: string
          description: "Time in HH:MM format"
        dayOfWeek:
          type: integer
          nullable: true
          description: "0=Sunday, 6=Saturday"
        dayOfMonth:
          type: integer
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - message
        - type
        - frequency
        - time
        - isActive
        - createdAt
    CreateNotificationInput:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [daily_review, weekly_summary, bill_reminder, budget_alert, custom]
        frequency:
          type: string
          enum: [daily, weekly, monthly, once]
        time:
          type: string
        dayOfWeek:
          type: integer
          nullable: true
        dayOfMonth:
          type: integer
          nullable: true
        isActive:
          type: boolean
      required:
        - title
        - message
        - type
        - frequency
        - time
    UpdateNotificationInput:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [daily_review, weekly_summary, bill_reminder, budget_alert, custom]
        frequency:
          type: string
          enum: [daily, weekly, monthly, once]
        time:
          type: string
        dayOfWeek:
          type: integer
          nullable: true
        dayOfMonth:
          type: integer
          nullable: true
        isActive:
          type: boolean
...